fixing bugs manual filtering and merging papers

+++ analysis/manual.py
@@ -1,6 +1,7 @@
 import pandas as pd
 from analysis import util
 import os
+from os.path import exists
 
 fr = 'utf-8'
 
@@ -10,39 +11,40 @@ def manual_filter_by_abstract(folder_name, next_file, search_date, step):
     papers_file = './papers/' + folder_name + '/' + str(search_date).replace('-', '_') + '/' + next_file
     file_name = './papers/' + folder_name + '/' + str(search_date).replace('-', '_') + '/' + str(step) + \
                 '_manually_filtered_by_abstract_papers.csv'
-    unknown_papers = 1
-    while unknown_papers > 0:
+    if exists(papers_file):
         to_check_papers = pd.read_csv(papers_file)
-        to_check_papers['title'] = to_check_papers['title'].str.lower()
-        to_check_papers = to_check_papers.drop_duplicates('title')
-        to_check_papers['abstract_norm'] = to_check_papers['abstract'].str.lower()
-        to_check_papers['abstract_norm'] = to_check_papers['abstract_norm'].str.replace(' ', '')
-        to_check_papers = to_check_papers.drop_duplicates('abstract_norm')
-        to_check_papers = to_check_papers.drop(columns=['abstract_norm'])
-        total_papers = len(to_check_papers)
         unknown_papers = len(to_check_papers.loc[to_check_papers['status'] == 'unknown'])
-        included_papers = len(to_check_papers.loc[to_check_papers['status'] == 'included'])
-        excluded_papers = len(to_check_papers.loc[to_check_papers['status'] == 'not included'])
-        progress = round(((total_papers - unknown_papers) / total_papers) * 100, 2)
-        print('::: Progress --> ' + str(progress) + '% :::')
-        print(' ::: Included (' + str(included_papers) + ') ::: Excluded(' + str(excluded_papers) + ') ::: Unknown('
-              + str(unknown_papers) + ') :::')
-        if len(to_check_papers.loc[to_check_papers['status'] == 'unknown']) > 0:
-            to_check_paper = to_check_papers.loc[to_check_papers['status'] == 'unknown'].sample()
-            print_paper_info(to_check_paper, file_name)
-            included, algorithm_type, training_schema, algorithm_goal, architecture = ask_manual_input()
-            paper_id = to_check_paper['id'].values[0]
-            if included == 'included':
-                paper_dict = {'id': (included_papers + 1), 'status': 'unknown', 'doi': to_check_paper['doi'],
-                              'publisher': to_check_paper['publisher'], 'database': to_check_paper['database'],
-                              'query_name': to_check_paper['query_name'], 'query_value': to_check_paper['query_value'],
-                              'url': to_check_paper['url'], 'publication_date': to_check_paper['publication_date'],
-                              'title': to_check_paper['title'], 'abstract': to_check_paper['abstract']
-                              }
-                paper_df = pd.DataFrame.from_dict(paper_dict)
-                with open(file_name, 'a+', newline='', encoding=fr) as f:
-                    paper_df.to_csv(f, encoding=fr, index=False, header=f.tell() == 0)
-            update_semantic_filtered_papers(to_check_papers, papers_file, paper_id, included)
+        while unknown_papers > 0:
+            to_check_papers['title'] = to_check_papers['title'].str.lower()
+            to_check_papers = to_check_papers.drop_duplicates('title')
+            to_check_papers['abstract_norm'] = to_check_papers['abstract'].str.lower()
+            to_check_papers['abstract_norm'] = to_check_papers['abstract_norm'].str.replace(' ', '')
+            to_check_papers = to_check_papers.drop_duplicates('abstract_norm')
+            to_check_papers = to_check_papers.drop(columns=['abstract_norm'])
+            total_papers = len(to_check_papers)
+            unknown_papers = len(to_check_papers.loc[to_check_papers['status'] == 'unknown'])
+            included_papers = len(to_check_papers.loc[to_check_papers['status'] == 'included'])
+            excluded_papers = len(to_check_papers.loc[to_check_papers['status'] == 'not included'])
+            progress = round(((total_papers - unknown_papers) / total_papers) * 100, 2)
+            print('::: Progress --> ' + str(progress) + '% :::')
+            print(' ::: Included (' + str(included_papers) + ') ::: Excluded(' + str(excluded_papers) + ') ::: Unknown('
+                  + str(unknown_papers) + ') :::')
+            if len(to_check_papers.loc[to_check_papers['status'] == 'unknown']) > 0:
+                to_check_paper = to_check_papers.loc[to_check_papers['status'] == 'unknown'].sample()
+                print_paper_info(to_check_paper, file_name)
+                included, algorithm_type, training_schema, algorithm_goal, architecture = ask_manual_input()
+                paper_id = to_check_paper['id'].values[0]
+                if included == 'included':
+                    paper_dict = {'id': (included_papers + 1), 'status': 'unknown', 'doi': to_check_paper['doi'],
+                                  'publisher': to_check_paper['publisher'], 'database': to_check_paper['database'],
+                                  'query_name': to_check_paper['query_name'], 'query_value': to_check_paper['query_value'],
+                                  'url': to_check_paper['url'], 'publication_date': to_check_paper['publication_date'],
+                                  'title': to_check_paper['title'], 'abstract': to_check_paper['abstract']
+                                  }
+                    paper_df = pd.DataFrame.from_dict(paper_dict)
+                    with open(file_name, 'a+', newline='', encoding=fr) as f:
+                        paper_df.to_csv(f, encoding=fr, index=False, header=f.tell() == 0)
+                update_semantic_filtered_papers(to_check_papers, papers_file, paper_id, included)
 
 
 def print_paper_info(to_check_paper, file_name):
@@ -92,34 +94,35 @@ def update_semantic_filtered_papers(to_check_papers, papers_file, paper_id, incl
 def manual_filter_by_full_text(folder_name, search_date, step):
     papers_file = './papers/' + folder_name + '/' + str(search_date).replace('-', '_') + '/' + str(step-1) + '_manually_filtered_by_abstract_papers.csv'
     file_name = './papers/' + folder_name + '/' + str(search_date).replace('-', '_') + '/' + str(step) + '_manually_filtered_by_full_text_papers.csv'
-    not_classified = 1
-    while not_classified > 0:
+    if exists(papers_file):
         filtered_by_abstract = pd.read_csv(papers_file)
-        total_papers = len(filtered_by_abstract)
         not_classified = len(filtered_by_abstract.loc[filtered_by_abstract['status'] == 'unknown'])
-        included_papers = len(filtered_by_abstract.loc[filtered_by_abstract['status'] == 'included'])
-        excluded = len(filtered_by_abstract.loc[filtered_by_abstract['status'] == 'excluded'])
-        progress = round(((total_papers - not_classified) / total_papers) * 100, 2)
-        print('::: Progress --> ' + str(progress) + '% :::')
-        print(' ::: Included Papers (' + str(included_papers) + ') ::: ')
-        print(' ::: Excluded (' + str(excluded) + ') ::: Not Classified(' + str(not_classified) + ') :::')
-        if len(filtered_by_abstract.loc[filtered_by_abstract['status'] == 'unknown']) > 0:
-            to_check_paper = filtered_by_abstract.loc[filtered_by_abstract['status'] == 'unknown'].sample()
-            print_paper_info_full_paper(to_check_paper, file_name)
-            t, title = ask_manual_input_full_paper()
-            if len(title) == 0:
-                title = to_check_paper['title']
-            paper_id = to_check_paper['id'].values[0]
-            if t == 'included':
-                paper_dict = {'id': (included_papers + 1), 'status': t, 'doi': to_check_paper['doi'],
-                              'publisher': to_check_paper['publisher'], 'database': to_check_paper['database'],
-                              'query_name': to_check_paper['query_name'], 'query_value': to_check_paper['query_value'],
-                              'url': to_check_paper['url'], 'publication_date': to_check_paper['publication_date'],
-                              'title': title, 'abstract': to_check_paper['abstract']}
-                paper_df = pd.DataFrame.from_dict(paper_dict)
-                with open(file_name, 'a+', newline='', encoding=fr) as f:
-                    paper_df.to_csv(f, encoding=fr, index=False, header=f.tell() == 0)
-            update_filtered_papers_by_abstract(filtered_by_abstract, papers_file, paper_id, t)
+        while not_classified > 0:
+            total_papers = len(filtered_by_abstract)
+            not_classified = len(filtered_by_abstract.loc[filtered_by_abstract['status'] == 'unknown'])
+            included_papers = len(filtered_by_abstract.loc[filtered_by_abstract['status'] == 'included'])
+            excluded = len(filtered_by_abstract.loc[filtered_by_abstract['status'] == 'excluded'])
+            progress = round(((total_papers - not_classified) / total_papers) * 100, 2)
+            print('::: Progress --> ' + str(progress) + '% :::')
+            print(' ::: Included Papers (' + str(included_papers) + ') ::: ')
+            print(' ::: Excluded (' + str(excluded) + ') ::: Not Classified(' + str(not_classified) + ') :::')
+            if len(filtered_by_abstract.loc[filtered_by_abstract['status'] == 'unknown']) > 0:
+                to_check_paper = filtered_by_abstract.loc[filtered_by_abstract['status'] == 'unknown'].sample()
+                print_paper_info_full_paper(to_check_paper, file_name)
+                t, title = ask_manual_input_full_paper()
+                if len(title) == 0:
+                    title = to_check_paper['title']
+                paper_id = to_check_paper['id'].values[0]
+                if t == 'included':
+                    paper_dict = {'id': (included_papers + 1), 'status': t, 'doi': to_check_paper['doi'],
+                                  'publisher': to_check_paper['publisher'], 'database': to_check_paper['database'],
+                                  'query_name': to_check_paper['query_name'], 'query_value': to_check_paper['query_value'],
+                                  'url': to_check_paper['url'], 'publication_date': to_check_paper['publication_date'],
+                                  'title': title, 'abstract': to_check_paper['abstract']}
+                    paper_df = pd.DataFrame.from_dict(paper_dict)
+                    with open(file_name, 'a+', newline='', encoding=fr) as f:
+                        paper_df.to_csv(f, encoding=fr, index=False, header=f.tell() == 0)
+                update_filtered_papers_by_abstract(filtered_by_abstract, papers_file, paper_id, t)
 
 
 def print_paper_info_full_paper(to_check_paper, file_name):
